Move `create_cycle_error` to `mod handle_cycle_error`
It's a more sensible home than `mod job`.
